\name{baselineARPLss-package}
\alias{baselineARPLss-package}
\alias{baselineARPLss}
\docType{package}
\title{
\packageTitle{baselineARPLss}
}
\description{
\packageDescription{baselineARPLss}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{baselineARPLss}
\packageIndices{baselineARPLss}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{baselineARPLss}

Maintainer: \packageMaintainer{baselineARPLss}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
